requirements.txt
werkzeug==2.3.7
Flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
mysql
pymysql
cryptography
mysql-connector-python
pytest